+++ src/animation/blinkanimation.cpp
@@ -8,7 +8,7 @@ namespace PixelMaestro {
 
 	void BlinkAnimation::map() {
 		// Alternate the Pixel between its normal color and off.
-		if (cycle_index_ == 0) {
+		if (frame_index_ == 0) {
 			if (orientation_ == Orientation::Vertical || orientation_ == Orientation::VerticalFlipped) {
 				for (uint16_t x = 0; x < section_.get_dimensions().x; x++) {
 					for (uint16_t y = 0; y < section_.get_dimensions().y; y++) {
@@ -35,6 +35,6 @@ namespace PixelMaestro {
 
 	void BlinkAnimation::update() {
 		map();
-		update_cycle(0, 2);
+		update_frame(0, 2);
 	}
 }