+++ src/animation/lightninganimation.cpp
@@ -34,7 +34,7 @@ namespace PixelMaestro {
 
 	void LightningAnimation::update() {
 		map();
-		update_cycle(0, palette_->get_num_colors());
+		update_frame(0, palette_->get_num_colors());
 	}
 
 	/**
@@ -84,7 +84,7 @@ namespace PixelMaestro {
 			}
 			cursor.x++;
 
-			set_map_color_index(x, cursor.y, cycle_index_ + bolt_num);
+			set_map_color_index(x, cursor.y, frame_index_ + bolt_num);
 
 			// Check to see if we should fork the bolt.
 			if (x < (uint16_t)section_.get_dimensions().x) {
@@ -147,7 +147,7 @@ namespace PixelMaestro {
 			}
 			cursor.y++;
 
-			set_map_color_index(cursor.x, y, cycle_index_ + bolt_num);
+			set_map_color_index(cursor.x, y, frame_index_ + bolt_num);
 
 			if (y < (uint16_t)section_.get_dimensions().y) {
 				uint8_t fork_roll = Utility::rand(UINT8_MAX);