+++ src/animation/plasmaanimation.cpp
@@ -54,6 +54,6 @@ namespace PixelMaestro {
 	}
 
 	void PlasmaAnimation::update() {
-		update_cycle(0, palette_->get_num_colors());
+		update_frame(0, palette_->get_num_colors());
 	}
 }