+++ src/animation/radialanimation.cpp
@@ -62,6 +62,6 @@ namespace PixelMaestro {
 	}
 
 	void RadialAnimation::update() {
-		update_cycle(0, palette_->get_num_colors());
+		update_frame(0, palette_->get_num_colors());
 	}
 }