+++ src/animation/waveanimation.cpp
@@ -32,6 +32,6 @@ namespace PixelMaestro {
 	}
 
 	void WaveAnimation::update() {
-		update_cycle(0, palette_->get_num_colors());
+		update_frame(0, palette_->get_num_colors());
 	}
 }